# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "API documentation for the News Aggregator project.",
        "contact": {
            "email": "manish011994@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Fetch articles with optional filters and pagination",
                "description": "Retrieve a list of articles based on various filters like keyword, category, source, date range, sorting, and ordering. Caching is used for optimization.",
                "operationId": "11aa4cc6561a7b94c56d1201d62513ee",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search keyword to filter articles by title or content",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter articles by category (array of category values)",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Filter articles by source (array of source values)",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Filter articles published after or on this date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Filter articles published before or on this date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort articles by a specific field. Possible values: title, published_at, category",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "title",
                                "published_at",
                                "category"
                            ]
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order for the articles. Possible values: asc, desc",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ArticleResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No articles found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No articles found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Fetch a single article by ID",
                "description": "Retrieve the details of a specific article by its ID. The response is cached for 10 minutes to optimize performance.",
                "operationId": "b3e529eab04efb5d305b07e2e58d9869",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the article to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article found."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ArticleResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [Article] with ID {id}."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ArticleResource": {
                "title": "Article",
                "description": "Schema representing an article",
                "properties": {
                    "id": {
                        "description": "ID of the article",
                        "type": "integer",
                        "example": 1
                    },
                    "author": {
                        "description": "Author of the article",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "title": {
                        "description": "Title of the article",
                        "type": "string",
                        "example": "Understanding OpenAPI"
                    },
                    "source": {
                        "description": "Source of the article",
                        "type": "string",
                        "example": "Tech News",
                        "nullable": true
                    },
                    "source_url": {
                        "description": "Source URL of the article",
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/article-source",
                        "nullable": true
                    },
                    "description": {
                        "description": "Description of the article",
                        "type": "string",
                        "example": "This article explains the basics of OpenAPI."
                    },
                    "keywords": {
                        "description": "Keywords associated with the article",
                        "type": "string",
                        "example": "OpenAPI, Swagger, API Documentation"
                    },
                    "category": {
                        "description": "Category of the article",
                        "type": "string",
                        "example": "Technology"
                    },
                    "image_url": {
                        "description": "Image URL associated with the article",
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/image.jpg",
                        "nullable": true
                    },
                    "content": {
                        "description": "Full content of the article",
                        "type": "string",
                        "example": "This is the content of the article..."
                    },
                    "published_at": {
                        "description": "Published date and time of the article",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-11T14:30:00Z"
                    },
                    "status": {
                        "description": "Status of the article",
                        "type": "string",
                        "enum": [
                            "draft",
                            "published",
                            "archived"
                        ],
                        "example": "published"
                    },
                    "created_at": {
                        "description": "Date and time when the article was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-10T10:00:00Z"
                    },
                    "updated_at": {
                        "description": "Date and time when the article was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-11T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Bearer token authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}